# order-service/src/main/resources/openapi/order-service.yml
openapi: 3.0.0
x-stoplight:
  id: g16hswxbggmiw
info:
  title: order-service
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  /api/address/addNewAddress:
    parameters: []
    post:
      summary: addNewAddress
      operationId: addNewAddress
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDto'
      tags:
        - address-controller
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressDto'
  /api/address/:
    get:
      summary: getAllAddress
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDto'
      operationId: getAllAddress
      tags:
        - address-controller
  '/api/address/{id}':
    parameters:
      - schema:
          type: integer
          format: int64
        name: id
        in: path
        required: true
    delete:
      summary: deleteAddress
      operationId: deleteAddress
      responses:
        '200':
          description: OK
      tags:
        - address-controller
      description: ''
  /api/order/:
    get:
      summary: getOrders
      tags:
        - order-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
      operationId: getOrders
  /api/order/addNewOrder:
    post:
      summary: addNewOrder
      operationId: addNewOrder
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
      tags:
        - order-controller
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDto'
    parameters: []
  '/api/order/find/{id}':
    parameters:
      - schema:
          type: integer
          format: int64
        name: id
        in: path
        required: true
    get:
      summary: findOrder
      tags:
        - order-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
      operationId: findOrder
      description: ''
      parameters: []
  /api/order/acceptOrder:
    parameters: []
    post:
      summary: acceptOrder
      operationId: acceptOrder
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
      parameters:
        - schema:
            type: integer
            format: int64
          in: query
          name: id
          required: true
        - schema:
            type: boolean
          in: query
          name: accept
          required: true
      tags:
        - order-controller
  /api/order/addItemToOrder:
    post:
      summary: addItemToOrder
      operationId: addItemToOrder
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
      parameters:
        - schema:
            type: integer
            format: int64
          in: query
          name: orderId
          required: true
        - schema:
            type: integer
            format: int64
          in: query
          name: productId
          required: true
        - schema:
            type: integer
          in: query
          name: piece
      tags:
        - order-controller
  '/api/order/{id}':
    parameters:
      - schema:
          type: integer
          format: int64
        name: id
        in: path
        required: true
    delete:
      summary: deleteOrder
      operationId: deleteOrder
      responses:
        '200':
          description: OK
      tags:
        - order-controller
  '/api/order/findByUser/{id}':
    parameters:
      - schema:
          type: integer
          format: int64
        name: id
        in: path
        required: true
    get:
      summary: findByUserId
      tags:
        - order-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
      operationId: findByUserId
      description: ''
components:
  schemas:
    AddressDto:
      title: AddressDto
      x-stoplight:
        id: cukxc2hw3e11j
      type: object
      properties:
        id:
          type: integer
          format: int64
        zipCode:
          type: string
        street:
          type: string
        houseNo:
          type: string
        cityName:
          type: string
    OrderDto:
      title: OrderDto
      x-stoplight:
        id: 98ci8le5pancz
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        priceSum:
          type: integer
        piecesSum:
          type: integer
        orderStatus:
          $ref: '#/components/schemas/OrderStatuses'
        address:
          $ref: '#/components/schemas/AddressDto'
        parcelno:
          type: string
        deliveryDate:
          type: string
          format: date
        orderItem:
          type: array
          items:
            $ref: '#/components/schemas/OrderedItemDto'
    OrderStatuses:
      title: OrderStatuses
      x-stoplight:
        id: rzsuwq8ff0lng
      type: string
    ItemDto:
      title: ItemDto
      x-stoplight:
        id: uhpglcb3f25na
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        productPrice:
          type: integer
        piece:
          type: integer
    CatalogDto:
      title: CatalogDto
      x-stoplight:
        id: moyogb2583xrn
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        productPrice:
          type: integer
        category:
          $ref: '#/components/schemas/CategoryDto'
    CategoryDto:
      title: CategoryDto
      x-stoplight:
        id: 5i0j2qs7p9udw
      type: object
      properties:
        id:
          type: integer
          format: int64
        categoryName:
          type: string
    OrderedItemDto:
      title: OrderedItemDto
      x-stoplight:
        id: hoc995ee8xuj6
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        productPrice:
          type: integer
        piece:
          type: integer